+++ build.gradle
@@ -32,7 +32,7 @@ dependencies {
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testImplementation 'org.springframework.security:spring-security-test'
 	implementation 'com.google.code.gson:gson:2.7' //test에서 사용
-	implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.0.2' //swagger
+	implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.3.0' //swagger
 	implementation 'org.springframework.boot:spring-boot-starter-validation'// valid 유효성 검사
 	implementation 'javax.validation:validation-api:2.0.1.Final'
 	implementation 'org.slf4j:slf4j-api:2.0.3' //logging
