+++ src/main/java/com/prolog/prologbackend/EmailTest/EmailService.java
@@ -1,17 +1,10 @@
-package com.prolog.prologbackend;
+package com.prolog.prologbackend.EmailTest;
 
-import com.prolog.prologbackend.DTO.EmailDTO;
-import jakarta.mail.Message;
-import jakarta.mail.MessagingException;
-import jakarta.mail.internet.InternetAddress;
-import jakarta.mail.internet.MimeMessage;
 import lombok.RequiredArgsConstructor;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
-
-import java.io.UnsupportedEncodingException;
 
 /**
  * Author : Kim
