+++ src/main/java/com/prolog/prologbackend/Exception/GlobalExceptionHandler.java
@@ -1,7 +1,5 @@
 package com.prolog.prologbackend.Exception;
 
-import javax.validation.ConstraintViolationException;
-
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
