+++ src/main/java/com/prolog/prologbackend/Members/DTO/UserEmail.java
@@ -1,4 +1,4 @@
-package com.prolog.prologbackend.DTO;
+package com.prolog.prologbackend.Members.DTO;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Email;
