+++ src/main/java/com/prolog/prologbackend/Project/Service/MypageProjectService.java
@@ -1,0 +1,8 @@
+package com.prolog.prologbackend.Project.Service;
+
+import com.prolog.prologbackend.Project.DTO.ProjectList;
+import com.prolog.prologbackend.Members.DTO.UserEmail;
+
+public interface MypageProjectService {
+    ProjectList getProjectList(UserEmail userEmail);
+}
