--- src/main/java/com/prolog/prologbackend/Repository/MypageRepository.java
@@ -1,8 +1,0 @@
-//package com.prolog.prologbackend.Repository;
-//
-//import com.prolog.prologbackend.Domain.members;
-//import org.springframework.data.jpa.repository.JpaRepository;
-//
-//public interface MypageRepository extends JpaRepository<members, String> {
-//    public members findByEmail(String email);
-//}
